+++ kroxylicious-sample/src/main/java/io/kroxylicious/sample/util/SampleFilterTransformer.java
@@ -61,7 +61,8 @@ public class SampleFilterTransformer {
 
             for (RecordBatch batch : records.batches()) {
                 for (Record batchRecord : batch) {
-                    newRecords.append(batchRecord.timestamp(), batchRecord.key(), transformRecord(batchRecord.value(), findValue, replacementValue),
+                    newRecords.appendWithOffset(batchRecord.offset(), batchRecord.timestamp(), batchRecord.key(),
+                            transformRecord(batchRecord.value(), findValue, replacementValue),
                             batchRecord.headers());
                 }
             }